feat(standard-schema): cleanup zod references
