+++ libs/ts-rest/express/package.json
@@ -8,12 +8,6 @@
   },
   "license": "MIT",
   "peerDependencies": {
-    "express": "^4.0.0",
-    "zod": "^3.24.0"
-  },
-  "peerDependenciesMeta": {
-    "zod": {
-      "optional": true
-    }
+    "express": "^4.0.0"
   }
 }
