+++ libs/ts-rest/serverless/package.json
@@ -13,8 +13,7 @@
   "peerDependencies": {
     "@azure/functions": "^4.0.0",
     "@types/aws-lambda": "^8.10.115",
-    "next": "^12.0.0 || ^13.0.0 || ^14.0.0",
-    "zod": "^3.24.0"
+    "next": "^12.0.0 || ^13.0.0 || ^14.0.0"
   },
   "peerDependenciesMeta": {
     "@azure/functions": {
@@ -24,9 +23,6 @@
       "optional": true
     },
     "next": {
-      "optional": true
-    },
-    "zod": {
       "optional": true
     }
   }
