+++ libs/ts-rest/solid-query/package.json
@@ -21,12 +21,6 @@
   },
   "license": "MIT",
   "peerDependencies": {
-    "@tanstack/solid-query": "^4.0.0",
-    "zod": "^3.24.0"
-  },
-  "peerDependenciesMeta": {
-    "zod": {
-      "optional": true
-    }
+    "@tanstack/solid-query": "^4.0.0"
   }
 }
